requirement.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.4.0
colorlog==2.10.0
coverage==4.5.1
Django==2.1.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-widget-tweaks==1.4.3
ez-setup==0.9
get==1.0.3
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
google-cloud-pubsub==0.38.0
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.15.0
idna==2.7
isort==4.3.4
Khayyam==3.0.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
Pillow==5.3.0
pluggy==0.8.0
post==1.0.2
protobuf==3.6.1
psq==0.7.0
psycopg2==2.7.5
public==2.0.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
pytest==3.9.1
pytest-cov==2.6.0
python-decouple==3.1
pytz==2018.5
query-string==1.0.2
request==1.0.2
requests==2.19.1
rsa==4.0
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
